C64.emu and keyboard input updates
* Imagine: Add support for passing the full meta key state in input events
* EmuFramework: Add support for passing meta key state to EmuSystem::handleInputAction()
* EmuFramework: Add EmuSystem::onVKeyboardShown()
* EmuFramework: Clean up option value read/write free functions
* EmuFramework: Add support for highlighting the shift key in VControllerKeyboard
* C64.emu: Update core VICE to 3.5
* C64.emu: Add additional key bindings for symbolic keys and combine both PC keyboard configs into one
* C64.emu: Update input handling to use keyboard_key_* functions and improve shifted key code mappings
* C64.emu: Add menu command to start emulation with an empty tape attached
* C64.emu: Add system option for selecting an external palette
* C64.emu: Replace Swap Joystick Ports option with Joystick Mode and add support for using joystick as keyboard cursor
* C64.emu: Replace cursor keys in keyboard overlay with left and up arrows

+++ C64.emu/src/vice/charset.h
@@ -29,9 +29,14 @@
 #define VICE_CHARSET_H
 
 #include "types.h"
+#include <stddef.h>
 
 #define a2p(c) charset_petconvstring(c, 0)
 #define p2a(c) charset_petconvstring(c, 1)
+
+
+/* TODO:    Fix these functions to use size_t for lenght, not int
+ */
 
 extern uint8_t *charset_petconvstring(uint8_t *c, int dir);
 extern uint8_t charset_p_toascii(uint8_t c, int cs);
@@ -45,7 +50,7 @@ extern void charset_petcii_to_screencode_line(const uint8_t *line, uint8_t **buf
                                               unsigned int *len);
 
 extern int charset_petscii_to_ucs(uint8_t c);
-extern int charset_ucs_to_utf8(uint8_t *out, int code, int len);
+extern int charset_ucs_to_utf8(uint8_t *out, int code, size_t len);
 
 #define CONVERT_TO_PETSCII 0
 #define CONVERT_TO_ASCII 1
